structure: Moved datagram stuff to it's own module
